requirements.txt
vllm==0.3.3
kaggle==1.6.0
faiss-cpu==1.8.0
numpy==1.26.4
torch==2.2.1
sentence-transformers==2.3.1
lean-dojo==0.1.2
python-dotenv==1.0.0